fix: rename HeliconeQueueProducer
