+++ valhalla/jawn/src/lib/proxy/ProxyForwarder.ts
@@ -1,6 +1,6 @@
 import { Response } from "node-fetch";
 import { Provider } from "@helicone-package/llm-mapper/types";
-import { HeliconeQueueProducer } from "../clients/HeliconeQuequeProducer";
+import { HeliconeQueueProducer } from "../clients/HeliconeQueueProducer";
 import { RequestWrapper } from "../requestWrapper/requestWrapper";
 import { getHeliconeAuthClient } from "../../packages/common/auth/server/AuthClientFactory";
 import { S3Client } from "../shared/db/s3Client";
